Logic moved to serializers
